refactor: consolidate push and workers into core/data
