+++ app/src/main/java/com/shrivatsav/monomail/MainActivity.kt
@@ -27,8 +27,8 @@ import com.shrivatsav.monomail.data.settings.FontScale
 import com.shrivatsav.monomail.data.settings.SettingsDataStore
 import com.shrivatsav.monomail.ui.navigation.NavGraph
 import com.shrivatsav.monomail.ui.theme.MonoMailTheme
-import com.shrivatsav.monomail.worker.EmailSyncWorker
-import com.shrivatsav.monomail.worker.GraphSubscriptionRenewalWorker
+import com.shrivatsav.monomail.core.data.worker.EmailSyncWorker
+import com.shrivatsav.monomail.core.data.worker.GraphSubscriptionRenewalWorker
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
