+++ app/src/main/java/com/shrivatsav/monomail/di/PushModule.kt
@@ -1,7 +1,7 @@
 package com.shrivatsav.monomail.di
 
-import com.shrivatsav.monomail.push.PushNotificationManager
-import com.shrivatsav.monomail.push.PushNotificationManagerImpl
+import com.shrivatsav.monomail.core.data.push.PushNotificationManager
+import com.shrivatsav.monomail.core.data.push.PushNotificationManagerImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
