+++ app/src/playstore/java/com/shrivatsav/monomail/push/FcmMessagingService.kt
@@ -10,7 +10,7 @@ import androidx.work.OutOfQuotaPolicy
 import androidx.work.WorkManager
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
-import com.shrivatsav.monomail.worker.EmailSyncWorker
+import com.shrivatsav.monomail.core.data.worker.EmailSyncWorker
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
